Split server functionality into its own module
Creates a new module 'server' which provides a function that runs either
an HTTP or HTTPS server as an future that can be passed to the async
runtime.
