+++ src/main.rs
@@ -1,22 +1,10 @@
-use async_stream::stream;
 use clap::{App, Arg};
-use core::task::{Context, Poll};
-use futures_util::{future::TryFutureExt, stream::Stream};
-use hyper::service::{make_service_fn, service_fn};
-use hyper::Server;
-use std::{net::SocketAddr, thread};
-use std::{io, process};
-use std::fs::File;
-use std::io::Read;
-use std::pin::Pin;
-use tokio::{net::{TcpListener, TcpStream}, runtime::Runtime, sync::mpsc::channel};
-use tokio_native_tls::{native_tls, TlsAcceptor, TlsStream};
-use openssl::pkcs12::Pkcs12;
-use openssl::pkey::PKey;
-use openssl::x509::X509;
+use std::{net::SocketAddr, process, thread};
+use tokio::{runtime::Runtime, sync::mpsc::channel};
 
 mod admission_review;
 mod api;
+mod server;
 mod wasm;
 mod wasm_fetcher;
 mod worker;
@@ -145,50 +133,11 @@ fn main() {
         worker_pool.run();
     });
 
-    macro_rules! mk_svc_fn {
-        ($tx:expr) => {
-            make_service_fn(|_conn| {
-                let svc_tx = $tx.clone();
-                async move {
-                    Ok::<_, hyper::Error>(service_fn(move |req| api::route(req, svc_tx.clone()))) }
-            });
-        };
-    }
-
-    rt.block_on( async {
-        if cert_file == "" {
-            let make_svc = mk_svc_fn!(api_tx);
-            let server = Server::bind(&addr).serve(make_svc);
-            println!("Started server on {}", addr);
-            if let Err(e) = server.await {
-                eprintln!("server error: {}", e);
-            }
-        } else {
-            let tls_acceptor = new_tls_acceptor(&cert_file, &key_file).unwrap();
-            let tcp = TcpListener::bind(&addr).await.unwrap();
-            let incoming_tls_stream = stream! {
-                loop {
-                    let (socket, _) = tcp.accept().await?;
-                    let stream = tls_acceptor.accept(socket).map_err(|e| {
-                        println!("[!] Voluntary server halt due to client-connection error...");
-                        error(format!("TLS Error: {:?}", e))
-                    });
-                    yield stream.await;
-                }
-            };
-
-            let make_svc = mk_svc_fn!(api_tx);
-            let server = Server::builder(HyperAcceptor {
-                acceptor: Box::pin(incoming_tls_stream),
-            }).serve(make_svc);
-
-            println!("Started server on {}", addr);
-            if let Err(e) = server.await {
-                eprintln!("server error: {}", e);
-            }
-        }
-    }
-    );
+    let tls_acceptor = match cert_file != "" {
+        true => Some(server::new_tls_acceptor(&cert_file, &key_file).unwrap()),
+        false => None,
+    };
+    rt.block_on(server::run_server(&addr, tls_acceptor, api_tx));
 
     wasm_thread.join().unwrap();
 }
@@ -196,40 +145,4 @@ fn main() {
 fn fatal_error(msg: String) {
     println!("{}", msg);
     process::exit(1);
-}
-
-fn error(err: String) -> io::Error {
-    io::Error::new(io::ErrorKind::Other, err)
-}
-
-struct HyperAcceptor<'a> {
-    acceptor: Pin<Box<dyn Stream<Item = Result<TlsStream<TcpStream>, io::Error>> + 'a>>,
-}
-
-impl hyper::server::accept::Accept for HyperAcceptor<'_> {
-    type Conn = TlsStream<TcpStream>;
-    type Error = io::Error;
-
-    fn poll_accept(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context,
-    ) -> Poll<Option<Result<Self::Conn, Self::Error>>> {
-        Pin::new(&mut self.acceptor).poll_next(cx)
-    }
-}
-
-fn new_tls_acceptor(cert_file: &str, key_file: &str) -> Result<TlsAcceptor, io::Error> {
-    let mut cert_file = File::open(cert_file)?;
-    let mut key_file = File::open(key_file)?;
-    let mut cert = vec![];
-    cert_file.read_to_end(&mut cert)?;
-    let cert = X509::from_pem(&cert)?;
-    let mut key = vec![];
-    key_file.read_to_end(&mut key)?;
-    let key = PKey::private_key_from_pem(&key)?;
-    let pkcs_cert = Pkcs12::builder()
-            .build("", "client", &key, &cert)?;
-    let identity = native_tls::Identity::from_pkcs12(&pkcs_cert.to_der()?, "").unwrap();
-    let tls_acceptor = native_tls::TlsAcceptor::new(identity).unwrap();
-    Ok(TlsAcceptor::from(tls_acceptor))
 }
